requirement.txt
selenium
webdriver_manager
uiautomator2
Appium-Python-Client